requirements.txt
Flask==2.3.3
flask-login==0.6.3
pymongo==4.5.0
Flask-PyMongo==2.3.0
dnspython==2.3.0
python-dotenv==1.0.0
bcrypt==4.0.1
flask-wtf==1.1.1
wtforms==3.0.1
gunicorn==20.1.0
